+++ sdk/storage/azure_storage_blob/src/clients/append_blob_client.rs
@@ -86,7 +86,7 @@ impl AppendBlobClient {
     /// # Arguments
     ///
     /// * `content_length` - Total length of the blob data to be uploaded.
-    /// * `options` - Optional configuration for the request. See [`AppendBlobClientCreateOptionsExt`](crate::models::AppendBlobClientCreateOptionsExt) for additional usage helpers.
+    /// * `options` - Optional configuration for the request.
     pub async fn create(
         &self,
         options: Option<AppendBlobClientCreateOptions<'_>>,
