+++ sdk/storage/azure_storage_blob/src/models/extensions.rs
@@ -2,23 +2,18 @@
 // Licensed under the MIT License.
 
 use crate::models::{
-    AppendBlobClientCreateOptions, BlobTag, BlobTags, BlockBlobClientPutBlobFromUrlOptions,
+    AppendBlobClientCreateOptions, BlobTag, BlobTags, BlockBlobClientUploadBlobFromUrlOptions,
     PageBlobClientCreateOptions,
 };
 use azure_core::error::ErrorKind;
 use std::collections::HashMap;
 
-/// Provides usage helpers for setting the `PageBlobClientCreateOptions` optional configurations.
-pub trait PageBlobClientCreateOptionsExt {
-    /// Augments the current options bag to only create if the Page blob does not already exist.
-    /// # Arguments
-    ///
-    /// * `self` - The options bag to be modified.
-    fn with_if_not_exists(self) -> Self;
-}
-
-impl PageBlobClientCreateOptionsExt for PageBlobClientCreateOptions<'_> {
-    fn with_if_not_exists(self) -> Self {
+/// Augments the current options bag to only create if the Page blob does not already exist.
+/// # Arguments
+///
+/// * `self` - The options bag to be modified.
+impl<'a> PageBlobClientCreateOptions<'a> {
+    pub fn with_if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -26,17 +21,12 @@ impl PageBlobClientCreateOptionsExt for PageBlobClientCreateOptions<'_> {
     }
 }
 
-/// Provides usage helpers for setting the `AppendBlobClientCreateOptions` optional configurations.
-pub trait AppendBlobClientCreateOptionsExt {
-    /// Augments the current options bag to only create if the Append blob does not already exist.
-    /// # Arguments
-    ///
-    /// * `self` - The options bag to be modified.
-    fn with_if_not_exists(self) -> Self;
-}
-
-impl AppendBlobClientCreateOptionsExt for AppendBlobClientCreateOptions<'_> {
-    fn with_if_not_exists(self) -> Self {
+/// Augments the current options bag to only create if the Append blob does not already exist.
+/// # Arguments
+///
+/// * `self` - The options bag to be modified.
+impl<'a> AppendBlobClientCreateOptions<'a> {
+    pub fn with_if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -44,16 +34,12 @@ impl AppendBlobClientCreateOptionsExt for AppendBlobClientCreateOptions<'_> {
     }
 }
 
-pub trait BlockBlobClientPutBlobFromUrlOptionsExt {
-    /// Augments the current options bag to only create if the Block blob does not already exist.
-    /// # Arguments
-    ///
-    /// * `self` - The options bag to be modified.
-    fn with_if_not_exists(self) -> Self;
-}
-
-impl BlockBlobClientPutBlobFromUrlOptionsExt for BlockBlobClientPutBlobFromUrlOptions<'_> {
-    fn with_if_not_exists(self) -> Self {
+/// Augments the current options bag to only create if the Block blob does not already exist.
+/// # Arguments
+///
+/// * `self` - The options bag to be modified.
+impl<'a> BlockBlobClientUploadBlobFromUrlOptions<'a> {
+    pub fn with_if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
