+++ sdk/storage/azure_storage_blob/src/models/mod.rs
@@ -40,9 +40,9 @@ pub use crate::generated::models::{
     BlobServiceProperties, BlobTag, BlobTags, BlobType, Block,
     BlockBlobClientCommitBlockListOptions, BlockBlobClientCommitBlockListResult,
     BlockBlobClientCommitBlockListResultHeaders, BlockBlobClientGetBlockListOptions,
-    BlockBlobClientPutBlobFromUrlOptions, BlockBlobClientPutBlobFromUrlResult,
-    BlockBlobClientPutBlobFromUrlResultHeaders, BlockBlobClientStageBlockOptions,
-    BlockBlobClientStageBlockResult, BlockBlobClientStageBlockResultHeaders,
+    BlockBlobClientStageBlockOptions, BlockBlobClientStageBlockResult,
+    BlockBlobClientStageBlockResultHeaders, BlockBlobClientUploadBlobFromUrlOptions,
+    BlockBlobClientUploadBlobFromUrlResult, BlockBlobClientUploadBlobFromUrlResultHeaders,
     BlockBlobClientUploadOptions, BlockBlobClientUploadResult, BlockBlobClientUploadResultHeaders,
     BlockList, BlockListType, BlockLookupList, ContainerItem, CopyStatus, CorsRule,
     DeleteSnapshotsOptionType, EncryptionAlgorithmType, FileShareTokenIntent,
